Main bot activation and main envenom functionality
Allow the user to enable/disable the envenom bot. Doing so clears the item list and any other tracking info, makes a request to the Invmon module to enable/disable itself, and produces a notification message on the announcements channel.

Added is_envenomable() method to Item class to determine whether an item is a weapon lacking the envenomed flag.

Modified the main envenom routine to check whether each item is envenomable. This allows that new additions to the item queue during an envenom operation are handled as expected, whereas using Envenom.filter_items would produce a separate table that could not grow while an operation ran.

TODO: Add a level check to Item:is_envenomable to ensure that the user's effective level is at least that of the item. Also probably safe to remove Envenom.filter_items.