New configs: channel, helmid, bagid
Channel config sets the channel on which bot notifications are broadcast. Default is gtell, as the main use of this plugin is assumed to be epic-related.

Helmid config defines the id (or keywords) of the user's Helm of True Sight, to be used to wear/remove it automatically as the envenom bot is enabled/disabled.

Bagid config defines the id (or keywords) of the container in which the user's helm should be stored, to be used to retrieve/store it automatically as the envenom bot is enabled/disabled.