Add check for running env operation
Previously, it was possible to initiate multiple envenom operations at once simply by entering the command multiple times. This could cause timeouts and ultimately serious breakage.

Added a new state variable, isEnvenoming, to the Envenom module. This is set to true when an envenom operation begins, and to false when it is completed. This way, only one operation should ever be able to run at a time, provided that any code that kicks off an operation checks this value first.