Allow cancellation of env operations
Added a new alias, aae cancel, to prematurely end an envenom operation if one is running. This uses the new isEnvenoming state variable to break out of the main envenom loop if it is ever set to false mid-operation.

Updated the envenom helpfile to reflect the new command. Also added colors to the same helpfile (oops).